+++ src/SharpLink.Abstractions/LogEvents.cs
@@ -77,5 +77,6 @@ public static class LogEvents
     {
         public const int UnknownOrTimedOutResponse = 6001;
         public const int BackgroundLoopUnhandledException = 6002;
+        public const int ConnectionAttemptFailed = 6101;
     }
 }